Migrate load selected buildings to function
Was 2 cached functions piped together which meant that the output of the
first cached function was being changed by the second and so streamlit
cache wouldn't operate as expected.  Now have 1 cached function which
will only rerun when the inputs to the function change - i.e. a user
selection change